renamed dispatcher and strategy
